+++ pysqlx_engine/_core/conn.py
@@ -2,7 +2,7 @@ from typing import Optional
 
 from .._core.parser import MyModel  # import necessary using _core to not subscribe default parser
 from .const import ISOLATION_LEVEL
-from .until import force_sync
+from .util import force_sync
 from .aconn import PySQLXEngine as _PySQLXEngine
